assign1/requirements.txt
requests
python-dotenv
pytz
pydantic
APScheduler
